+++ addons/nature/CfgPatchesUnits.hpp
@@ -1,197 +1,8 @@
 units[] = {
-    //Bushes
-    "Cane1",
-    "Cane2",
-    "Ficus_Bush_1",
-    "Ficus_Bush_2",
-    "Ficus_Bush_3",
-    "Oleander1",
-    "Oleander2",
-    "Oleander_White",
-    "Thistle_Thorn_Green_Bush",
-    "Vinifera",
-    "Opuntiaficusindica",
-    "Calochlaena",
-    "Cestrum",
-    "Colored_red",
-    "Colored_yellow",
-    "Cycas",
-    "FicusC2d_tanoa",
-    "Gardenia_dec_01",
-    "Gardenia_dec_02",
-    "Gardenia",
-    "Leucaena_shrub",
-    "NeriumO2d_tanoa",
-    "PiperMeth",
-    "Rhizophora_shrub",
-
-    //Rocks
-    "BluntRock_Apart",
-    "BluntRock_Monolith",
-    "BluntRock_Spike",
-    "BluntRock_WallH",
-    "BluntRock_WallV",
-    "BluntStone1",
-    "BluntStone2",
-    "BluntStone3",
-    "BluntStone_Erosion",
-
-    //Trees
-    "Paper_Mulberry",
-    "Ficus_1",
-    "Ficus_2",
-    "Fraxinus",
-    "Olive_1",
-    "Olive_2",
-    "Palm_1",
-    "Palm_2",
-    "Pine_1",
-    "Pine_2",
-    "Pine_3",
-    "Pine_4",
-    "Poplar_Dead",
-    "Poplar",
-    "Oak",
-    "Fallen_Branch1",
-    "Fallen_Branch2",
-    "Fallen_Branch3",
-    "Branch_Big",
-    "Cupressus_1",
-    "Cupressus_2",
-    "Ficus_3",
-    "Phoenixrupicola_1",
-    "Phoenixrupicola_2",
-    "Agathis_tall",
-    "Agathis_wide",
-    "Albizia",
-    "Cocos_bend",
-    "Cocos_small",
-    "Cocos_tall",
-    "CocosNucifera2s_small",
-    "CocosNucifera3s_bend",
-    "CocosNucifera3s_tall",
-    "Cyathea",
-    "Ficus_big",
-    "Ficus_medium",
-    "Ficus_small",
-    "Inocarpus",
-    "Leucaena",
-    "Millettia",
-    "Millettia_plantation",
-    "Palaquium",
-    "Pritchardia",
-    "Rhizophora",
-    "FallenBranch",
-    "FallenTrunk_branches",
-    "FallenTrunk_clear",
-    "FallenTrunk_roots",
-    "TreeStump_cut_large",
-    "TreeStump_cut_small",
-    "TreeStump_natural_large",
-    "TreeStump_natural_small",
-    "Banana",
-    "Banana_slim",
-    "Banana_wild",
-
-    //Plants
-    "Carduus",
-    "Crack_Clutter1",
-    "Crack_Clutter2",
-    "Crack_Clutter3",
-    "Crack_Clutter4",
-    "Flower_Medium_Blue",
-    "Flower_Medium_White",
-    "Flower_Cakile",
-    "Flower_Low_Yellow1",
-    "Flower_Low_Yellow2",
-    "Flower_Medium_White1",
-    "Flower_Medium_White2",
-    "Flower_Tansy",
-    "Grass_High_Dry",
-    "Grass_High_Green",
-    "Grass_Low_Sharp",
-    "Grass_Bunch_Small",
-    "Grass_Dry",
-    "Grass_Green",
-    "Grass_Tall_Dead",
-    "Grass_Tuft_Dry",
-    "Grass_Crooked",
-    "Grass_Tuft_Dead",
-    "Grass_Desert_GroupHard",
-    "Grass_Green_GroupHard",
-    "Grass_Green_GroupSoft",
-    "Grass_Long_Dry",
-    "Grass_Tall",
-    "Plant_chicory",
-    "Plant_Dock",
-    "Plant_Green_Small",
-    "Plant_Green_Bunch",
-    "Plant_Mullein",
-    "Rubble_Clutter1",
-    "Rubble_Clutter2",
-    "Rubble_Clutter3",
-    "Rubble_Clutter4",
-    "Rubble_Clutter5",
-    "Rubble_Clutter6",
-    "Flower_White_Small",
-    "Str_Grass_Dry",
-    "Str_WeedBrown_Tall",
-    "Str_WeedGreen_Tall",
-    "Plant_Germander_Group",
-    "Plant_Green_Shrub",
-    "Thistle_Purple_Small",
-    "Thistle_Yellow_Small",
-    "Thistle_Yellow_Shrub",
-    "Thron_Gray",
-    "Thron_Green",
-    "Thron_Khaki",
-    "Thistle_High",
-    "Thistle_High_Dead",
-    "Thistle_Small",
-    "Thistle_Small_GreenYellow",
-    "Thistle_Small_Yellow",
-    "Thistle_Thorn_Brown",
-    "Thistle_Thorn_Desert",
-    "Thistle_Thorn_Gray",
-    "Thistle_Thorn_Green",
-    "Thistle_Thorn_Green_Bush",
-    "Reeds",
-    "Ginger_20",
-    "Ginger_3",
-    "Ginger_9",
-    "Sugarcane_mature_row",
-    "Sugarcane_sapling_row",
-    "Sugarcane_sapling_single",
-    "Cacao_ripe",
-    "LianaTangle_narrow",
-    "LianaTangle_wide",
-    "Grass_leaf_big",
-    "Grass_Leaves_coltsfoot",
-    "Grass_mimosa",
-    "Grass_nettle",
-    "Grass_short_bunch",
-    "Grass_short_leaf",
-    "Grass_short_mimosa",
-    "Grass_short_nettle",
-    "Grass_short_small",
-    "Grass_Tropic",
-    "GrassBunch_HI",
-    "GrassBunch_LO",
-    "forest_BiglLeaves",
-    "forest_BiglLeaves2",
-    "forest_fern",
-    "forest_roots",
-    "forest_violet_leaves",
-    "forest_violet_leaves2",
-    "forest_violet_single",
-    "volcano_Grass",
-    "volcano_GrassMix",
-    "volcano_GrassMix2",
-    "red_dirt_leaves",
-    "red_dirt_leaves_w3",
-    "red_dirt_mimosa",
-    "red_dirt_short_Grass",
-    "red_dirt_sparse_Grass",
-    "red_dirt_sparse_Grass2",
-    "red_dirt_sparse_Grass3"
+    #include "cfgpatchesunits\Classnames_Bush.hpp"
+    #include "cfgpatchesunits\Classnames_Clutter.hpp"
+    #include "cfgpatchesunits\Classnames_Grass.hpp"
+    #include "cfgpatchesunits\Classnames_Plant.hpp"
+    #include "cfgpatchesunits\Classnames_Rock.hpp"
+    #include "cfgpatchesunits\Classnames_Tree.hpp"
 };
