+++ addons/nature/cfgpatchesunits/Classnames_Bush.hpp
@@ -1,0 +1,28 @@
+//ARMA Vanilla
+"Cane1",
+"Cane2",
+"Ficus_Bush_1",
+"Ficus_Bush_2",
+"Ficus_Bush_3",
+"Oleander1",
+"Oleander2",
+"Oleander_White",
+
+//ARMA Malden
+"Vinifera",
+"Opuntiaficusindica",
+
+//ARMA APEX
+"Calochlaena",
+"Cestrum",
+"Colored_red",
+"Colored_yellow",
+"Cycas",
+"FicusC2d_tanoa",
+"Gardenia_dec_01",
+"Gardenia_dec_02",
+"Gardenia",
+"Leucaena_shrub",
+"NeriumO2d_tanoa",
+"PiperMeth",
+"Rhizophora_shrub",
