+++ addons/nature/cfgpatchesunits/Classnames_Clutter.hpp
@@ -1,0 +1,12 @@
+//ARMA Vanilla
+"Crack_Clutter1",
+"Crack_Clutter2",
+"Crack_Clutter3",
+"Crack_Clutter4",
+
+"Rubble_Clutter1",
+"Rubble_Clutter2",
+"Rubble_Clutter3",
+"Rubble_Clutter4",
+"Rubble_Clutter5",
+"Rubble_Clutter6",
