+++ addons/nature/cfgpatchesunits/Classnames_Plant.hpp
@@ -1,0 +1,66 @@
+//ARMA Vanilla
+"Carduus",
+"Flower_Medium_Blue",
+"Flower_Medium_White",
+"Flower_Cakile",
+"Flower_Low_Yellow1",
+"Flower_Low_Yellow2",
+"Flower_Medium_White1",
+"Flower_Medium_White2",
+"Flower_Tansy",
+"Plant_chicory",
+"Plant_Dock",
+"Plant_Green_Small",
+"Plant_Green_Bunch",
+"Plant_Mullein",
+"Flower_White_Small",
+"Plant_Germander_Group",
+"Plant_Green_Shrub",
+"Thistle_Purple_Small",
+"Thistle_Yellow_Small",
+"Thistle_Yellow_Shrub",
+"Thron_Gray",
+"Thron_Green",
+"Thron_Khaki",
+"Thistle_High",
+"Thistle_High_Dead",
+"Thistle_Small",
+"Thistle_Small_GreenYellow",
+"Thistle_Small_Yellow",
+"Thistle_Thorn_Brown",
+"Thistle_Thorn_Desert",
+"Thistle_Thorn_Gray",
+"Thistle_Thorn_Green",
+"Thistle_Thorn_Green_Bush",
+"Reeds",
+
+//ARMA APEX
+"Ginger_20",
+"Ginger_3",
+"Ginger_9",
+"Sugarcane_mature_row",
+"Sugarcane_sapling_row",
+"Sugarcane_sapling_single",
+
+"LianaTangle_narrow",
+"LianaTangle_wide",
+
+"Grass_Leaves_coltsfoot",
+"Grass_mimosa",
+"Grass_nettle",
+"Grass_short_leaf",
+"Grass_short_mimosa",
+"Grass_short_nettle",
+"Grass_Tropic",
+
+"forest_BiglLeaves",
+"forest_BiglLeaves2",
+"forest_fern",
+"forest_roots",
+"forest_violet_leaves",
+"forest_violet_leaves2",
+"forest_violet_single",
+
+"red_dirt_leaves",
+"red_dirt_leaves_w3",
+"red_dirt_mimosa",
