+++ addons/nature/cfgpatchesunits/Classnames_Rock.hpp
@@ -1,0 +1,10 @@
+//ARMA Vanilla
+"BluntRock_Apart",
+"BluntRock_Monolith",
+"BluntRock_Spike",
+"BluntRock_WallH",
+"BluntRock_WallV",
+"BluntStone1",
+"BluntStone2",
+"BluntStone3",
+"BluntStone_Erosion",
