+++ addons/nature/cfgpatchesunits/Classnames_Tree.hpp
@@ -1,0 +1,64 @@
+//ARMA Vanilla
+"Paper_Mulberry",
+"Ficus_1",
+"Ficus_2",
+"Fraxinus",
+"Olive_1",
+"Olive_2",
+"Palm_1",
+"Palm_2",
+"Pine_1",
+"Pine_2",
+"Pine_3",
+"Pine_4",
+"Poplar_Dead",
+"Poplar",
+"Oak",
+"Fallen_Branch1",
+"Fallen_Branch2",
+"Fallen_Branch3",
+"Branch_Big",
+
+//ARMA Malden
+"Cupressus_1",
+"Cupressus_2",
+"Ficus_3",
+"Phoenixrupicola_1",
+"Phoenixrupicola_2",
+
+//ARMA APEX
+"Agathis_tall",
+"Agathis_wide",
+"Albizia",
+"Cocos_bend",
+"Cocos_small",
+"Cocos_tall",
+"CocosNucifera2s_small",
+"CocosNucifera3s_bend",
+"CocosNucifera3s_tall",
+"Cyathea",
+"Ficus_big",
+"Ficus_medium",
+"Ficus_small",
+"Inocarpus",
+"Leucaena",
+"Millettia",
+"Millettia_plantation",
+"Palaquium",
+"Pritchardia",
+"Rhizophora",
+
+"Cacao_ripe",
+
+"FallenBranch",
+"FallenTrunk_branches",
+"FallenTrunk_clear",
+"FallenTrunk_roots",
+"TreeStump_cut_large",
+"TreeStump_cut_small",
+"TreeStump_natural_large",
+"TreeStump_natural_small",
+
+"Banana",
+"Banana_slim",
+"Banana_wild"
