+++ addons/nature/cfgvehicles/CfgVehicles_Clutter.hpp
@@ -3,6 +3,7 @@ CLUTTER(Crack_Clutter1,A3\plants_f\Clutter\c_Crack_Clutter1.p3d);
 CLUTTER(Crack_Clutter2,A3\plants_f\Clutter\c_Crack_Clutter2.p3d);
 CLUTTER(Crack_Clutter3,A3\plants_f\Clutter\c_Crack_Clutter3.p3d);
 CLUTTER(Crack_Clutter4,A3\plants_f\Clutter\c_Crack_Clutter4.p3d);
+
 CLUTTER(Rubble_Clutter1,A3\plants_f\Clutter\c_Rubble_Clutter1.p3d);
 CLUTTER(Rubble_Clutter2,A3\plants_f\Clutter\c_Rubble_Clutter2.p3d);
 CLUTTER(Rubble_Clutter3,A3\plants_f\Clutter\c_Rubble_Clutter3.p3d);
