+++ addons/nature/cfgvehicles/CfgVehicles_Plant.hpp
