+++ addons/nature/cfgvehicles/CfgVehicles_Rock.hpp
@@ -1,14 +1,12 @@
+//ARMA Vanilla
 ROCK(BluntRock_Apart,A3\rocks_f\Blunt\BluntRock_apart.p3d);
 ROCK(BluntRock_Monolith,A3\rocks_f\Blunt\BluntRock_monolith.p3d);
 ROCK(BluntRock_Spike,A3\rocks_f\Blunt\BluntRock_spike.p3d);
 ROCK(BluntRock_WallH,A3\rocks_f\Blunt\BluntRock_wallH.p3d);
 ROCK(BluntRock_WallV,A3\rocks_f\Blunt\BluntRock_wallV.p3d);
 ROCK(BluntStone1,A3\rocks_f\Blunt\BluntStone_01.p3d);
-//ROCK(BluntStone1_LC,A3\rocks_f\Blunt\BluntStone_01_LC.p3d);
 ROCK(BluntStone2,A3\rocks_f\Blunt\BluntStone_02.p3d);
-//ROCK(BluntStone2_LC,A3\rocks_f\Blunt\BluntStone_02_LC.p3d);
 ROCK(BluntStone3,A3\rocks_f\Blunt\BluntStone_03.p3d);
-//ROCK(BluntStone3_LC,A3\rocks_f\Blunt\BluntStone_03_LC.p3d);
 ROCK(BluntStone_Erosion,A3\rocks_f\Blunt\BluntStones_erosion.p3d);
 
 //ROCK(NAME,MODELPATH);
